feat: create readme parser
Created src\util\loadReadmes.js for extracting info out of README files
and storing them in JSON files.

Installed the markdown-it package to make getting README text easier.

Added readmes and repositories folders to .gitignore, because it is not
necessary to keep old files and folders once the website has been built.

!Contains temp code that should be removed in the future.