feat: add about section HTML
Placed the text for the "about" section in a text file in a new folder
to better separate concerns.  Originally, it was done to allow updating
the text file without having to make a commit, but updates will be rare
and there was a chance of losing the "about" text if it wasn't
committed.

There is some initial CSS styling applied.