fix: make img path relative to project root
In the extracted README data in the JSON files in /readmes folder, the
img src paths were relative to repository README.md files, which made
it difficult to use when rendering images in a React component, since
the repository name was not being stored.  Although it is possible to
figure out the path when importing the README JSON files, it might be
adding unrelated code in places where the code does not belong.
Alternatively, the path to image files from the project root is now
saved instead.

In the getImgs function, the img src is now the path from the project
root.

Functions parseFirstSection, getImgs, and cleanFirstSection have become
nested functions in the parseReadme function, to prevent needlessly
passing the repository name between intermediary functions.  The
functions also are only used when parsing the README and no where else.

Also changed from using relative imports to absolute imports to
maintain consistency in file/folder paths.

Also removed exporting the parseReadme function, because having both a
call to loadReadmes() and the export does not make sense.  If
loadReadmes.js is imported, it would cause loadReadmes to be
unintentionally called.  There is also no current need to export
parseReadme.