fix: refine selection of project images
In the front-end, when displaying project images on the main webpage,
there was no way of selecting the best image to represent a project.
The home page image may not necessarily give a good idea of what the
project looks like.  When viewing the details of a project, the images
may be ordered differently as well.  So being able to choose which
single image to be the primary image and which images to be shown in
order in the details webpage of a project is necessary.

A new file, indexImages.json, will need to be created in the
readme_files directory for each repository.  It will need to contain an
Object, with keys "primary" and "all".  The "primary" key will need to
contain an image Object.  The "all" key will need to contain a list of
image Objects.  An image Object contains "src" and "alt" keys, which
will be used in an HTML <img> tag.
Example:
{
  "primary": { "src": "home_page.jpg", "alt": "Home page" },
  "all": [
    { "src": "home_page.jpg", "alt": "Home page" },
    { "src": "details_page.jpg", "alt": "Details page" },
    { "src": "search_results.jpg", "alt": "Search Results" }
  ]
}

For the code change, the old code has been entirely removed, because
searching for images in the README is no longer necessary.  Instead,
the indexImages.json file, which will contain the info of the images to
use, is loaded and modified with the correct path to the images from
the project root.

Since getting the images is no longer related to the first section of a
README, getImgs is now called in the parseReadme function.
parseFirstSection's code and JSDoc has been updated accordingly.