feat: add projects HTML
Installed html-react-parser package to parse the raw Strings of the
sections in the readme JSONs.

Implemented the Projects and ProjectCard React components to display
the projects section and basic info of each project.

The code to import all readme JSONs is in Projects, which should be
fine for now, because it is in a useEffect block and only executed on
initial mount.

A little bit of CSS styling has been applied, but it's incomplete.