feat: add project details page HTML
Created ProjectPage React component to display the details of a
project.  It parses the readme JSONs.

In the App component, included the rendering of ProjectPage.  Routes
were not used, because I do not want users to directly visit other
parts of the portfolio website and get a "URL not found" error.  Since
there is no routing, the logic to "open" project pages is done through
state with callback functions, that change the state, being passed down
to child components through a React context.

In the ProjectCard component, the card is made clickable so that the
website can "open" pages to project details.

Installed @mui/icons-material (Material UI icons) to use a back arrow
icon on ProjectPage.

Initial and partial CSS styling have been done.