feat: add cardboard box to About section
Updated About component to with some styling.

Created a cardboard box SVG by using ChatGPT to generate an initial
drawing and using Inkscape to retrace the drawing and saving it as a
SVG.  In the implementation in the About component, the box uses height
and width values that prevents it from extending the viewport outside
the website's overall container (prevents horizontal scrolling).

Created a CSS block in appCss.js to center elements, with the new class
"flex-center".