feat: add CSS styling to Projects section
In the Projects component, added a Grid layout so that there can be one
project per row on small screens and two projects per row on big
screens.

In the ProjectCard component, added the cardboard box background to
each project card.  Also removed the width setting for CardMedia so
that the image can fill the width of the card.

Note that the project cards are small, so the brief description of each
project has to be short.