fix: don't apply margins to ProjectPage sections
In appCss.js, the CSS property for applying a vertical margin for "main
> section" was also being applied to sections in the project details
page, which was too much space for that page.  To fix that, a :not
pseudo-selector is used, but since it increased the specificity of the
selector, it also caused other parts of the website to break, because
the CSS styling implementations were incomplete.

An additional fix had to be performed for the Skills component.  The
top returned component is changed from Box to Container, so that the
width of the skills section takes up the width of the viewport.  The
unordered list was also missing the removal of the bullets.

An additional fix is made to the App component.  The main part of the
website, that contains the sections, had to have a flexbox applied to
center everything horizontally.