feat: parse Tech Stack raw HTML String
Created a function, parseTechStack, to further parse the Tech Stack raw
HTML into an Object.  This will allow flexibility in putting the tech
stack sections and items into what ever HTML element or Material UI
component that I need.

Also corrected JSDoc for getImgs, because it was no longer an
asynchronous function and returning a Promise.