feat: add CSS styling to and rewrote project page
Added a reusable HorizontalLine component to separate sections.

Added a reusable H2 component to underline the headings.

Abstracted out the different parts of the project details page into
functions, because the return statement was too cluttered.

In the renderImages function, placed the images into a Stack component
so that the image alternative texts are stacked vertically.

In the renderFirstSection function, reverted making deployed links to
have display type of block, towards putting them in paragraph elements,
because the links were clickable beyond the expected widths.

Created a separate renderTechStack function specifically for the Tech
Stack section, to put the lists of technologies into Grid components.

In the renderRestOfSections function, changed the Array function from
map to reduce, so that horizontal lines could be inserted.  Also
excluded processing the Tech Stack section.

In src\App.jsx, moved the NavBar component from always being present to
only showing up on the home page, because links to the different parts
of the home page were not working when on the project details page.  It
is also done, because the original intention was to have a modal to
display the project details and have users focus entirely on it.  The
project details page was supposed to be something that was closeable.