feat: click logo to redirect to home
The click handling, for clicking on the logo in the NavBar, was a hack
to go back to the home page.  This hack was done by changing a state in
the App component.  A better way of going back to the home page and
starting over would be to make the logo an anchor link.