feat: add SVG of cardboard box side view
Created a SVG of a cardboard box from the side, so it can be used as
the hover background for links in the navbar.

Also changed the metadata document names of the other SVGs and the file
name of the first cardboard box SVG for clarity.  Updated file names in
code, accordingly.