feat: add images for nav links
Created a side view of a conveyor roller in SVG and placed that under
the navigation links.

Displayed a side view of a cardboard box on hover of links.
Consequently, in the NavLink component, removed the background color on
hover.

Increased toolbar height, along with scrollMarginTop CSS property for
main sections, to accommodate images.

For theme, changed primary color to blue grey to avoid camouflaging
conveyor roller.