feat: add CSS styling for main section titles
Placed the title headings for each section in Paper components so that
they are readable.

Refactored common code for each section into MainSectionsContainer,
MainSectionsHeading1, and MainSectionsHeading2 React components.  There
are separate components for headings 1 and 2, because they might have
more unique CSS styles for each of them in the future.

In appCss.js, the margin for each section has been moved to
MainSectionsContainer.  Since the headings are now in Paper components,
the bottom margins are no longer needed for headings 1 and 2.