feat: add CSS styling to project page
Added more styling to the project page, because the original
implementation was inadequate.  Used Material UI Typography components
for headings instead of raw HTML tags.

In the ProjectPage component, removed "justifyContent" and "alignItems"
CSS properties from the top-level component, because it was causing the
images to not fill the screen width.  Consequently, components for
sections and HorizontalLine had to have a Container inside a Box, so
that it is centered properly on the page.

In the ProjectTechStack component, the Grid size breakpoints had to be
changed, because text inside of it was overflowing for small screen
sizes.

Rewrote H2 component to maintain consistency in using Typography to
stylize text.

Added a custom secondary dark color, because the default one wasn't
dark enough.

Applied theme colors to text and background.