fix: reduce heading size on project page
The heading sizes on the project page were too large on small screens.
The visual rendering of the heading elements have been changed to be
the next smaller heading on small screens.  A new file,
projectPageHeadingStyles.js, is created to store this configuration for
all headings on the project page.