feat: make mobile version of NavBar
Placed the navigation links into a collapsible list for small screens.
Note that currently the cardboard box line is removed for small screens
due to time.  Also broke up the code for NavBar into variables and a
React component called Logo so that they are more readable.

Applied theme styles for text color.