fix: expand About box card for small screens
When the screen width is really small, the text in the cardboard box
overflows out of the card.  This is now corrected by setting min and
max height and width, instead of exact values.

Box components have also been replaced by Card components, because the
Card components better represent what is on the screen.