fix: reduce heading size on home page
The heading sizes on the home page were too large on small screens.
The visual rendering of the heading elements have been changed to be
the next smaller heading on small screens by using the configuration
file created earlier, in
src\components\ProjectPage\components\projectPageHeadingStyles.js.

The projectPageHeadingStyles file is moved to a common parent directory
and is renamed, because the homepage uses it now.  The path is now
src\constants.js.