fix: improve project page headings and dividers
Replaced the custom-built HorizontalLine component with Material UI's
Divider.  The Divider had to be within a div element, because it was
extending to its full length while assigned the middle variant.  I
assume it is because its container cannot be a flexbox.

Removed underline, because it did not look good.  Also applied theme
color to Tech Stack section's categories, even though it looks the same
visually.