fix: update loadReadmes.js for more repos
With the addition of more varied repository READMEs and images, the
loadReadmes.js file had to adapt to READMEs with the Tech Stack section
being the last section and READMEs with no images.

The parseReadme function was not saving the Tech Stack section, because
it was only saving section text when encountering a <h2> tag.  Since
some repositories had Tech Stack last, there would not be a <h2> tag
after it, which then meant that Tech Stack would not be saved.  This
was fixed by also storing section text when the line is an empty
String, which only occurs at the end of a file.

The getImgs function was throwing an Error, because there would not be
an indexImages.json file for READMEs without images, such as for
API-only projects.  This was fixed by simply returning an empty Object
if an Error was thrown.  Returning an empty Object was the simplest
solution and would not require changing the output README JSON data
structure too much.