fix: adjust project page layout
On the project-specific page, the Overview section had its text
centered.  If there are multiple paragraphs of texts, in <p> tags, then
this caused the paragraphs to be right next to each other and made it
difficult to read.  In ProjectRestOfSections.js, CSS styling had to be
added specifically to the Overview section in order to fix this.

For the Tech Stack section of the project page, there was empty space
at where a third category would be when the list of tech stack
categories was two.  This was fixed by not having the grid size be an
exact number at medium size, in ProjectTechStack.jsx.