feat: sort projects in descending on homepage
On the homepage, the projects listed should be sorted with the most
recent project first.

In order to accomplish this, a new file called metadata.json had to be
created, containing the creation date, for each project when
downloading from GitHub, because the creation date did not belong in
the README markdown file nor in the images/readme_files directory.  The
creation date did not exist further down the process line, because the
data has been trimmed.

The creation date had to then be transferred to the project JSON files
in the readmes directory during the conversion of the raw README
markdown files to JSON files.

The creation dates could then finally be used in the frontend in the
useEffect block of the Projects component, where it is then used to
sort the projects.  The sorting had to be done in the useEffect block,
because there was no way of sorting the files in the readmes directory
without changing the file names.

Also in the useEffect block of the Projects component, the creating and
calling of the asynchronous function has been simplified into an IIFE.