fix: move images into public directory
In the built, production-grade website, the images were not being
loaded, because Vite was not able to "see" them.  The images were not
in the "public" directory nor being imported; they were directly
accessed by file path, which worked in the development environment but
not in the production environment.

The brand/tech icons and background images are now placed in the
"public" directory.  Only those are placed and not the project images,
because those will always be used, regardless of website content.  They
also were not required for compiling.  In other words, they were simply
placed in CSS to display images.