fix: properly import image into ProjectCard
Since images can not be directly accessed by simply providing the file
path to an img or HTML element, the ProjectCard component has been
updated to use useEffect to load the project images.  useEffect is
needed, because using Vite's unique "import.meta.glob" feature would be
external to React.

Also made the CardMedia component, in the return of ProjectCard, a flex
container, so that the alt image text is centered.