fix: properly import images into ProjectImages
Images can not be directly accessed by simply providing the file path
to an img or HTML element, or else Vite would not save the images in
the production build.

For the images on the project-specific page, a subcomponent called
ProjectImage is created to help retrieve the individual images.
Creating this subcomponent is a simpler solution than creating an Array
of the retrieved images and associating the items in the images
variable with the items in the Array within ProjectImages.  It is also
the most straight-forward solution, as each image can be considered to
be its own component.

ProjectImage is within ProjectImages.jsx in order to simplify file
structure.  It is also only going to be used by ProjectImages.

In ProjectImage, useEffect is needed, because using Vite's unique
"import.meta.glob" feature would be considered external to React.