fix: run missing actions for GitHub Actions
In the static.yml for GitHub Actions, the steps relating to building
has now been separated from deployment steps to make the workflow more
readable.

The workflow was missing the downloading and parsing of repository
READMEs and README files.  Two new steps, "Download GitHub Repos" and
"Parse READMEs", have now been added to the Build phase to remediate
that.