fix: make paths relative within components
The images in the "public" directory were not being loaded on the
published website.  GitHub needs the paths to the files in the "public"
directory to be relative.

For the Logo component, the link also had an absolute path, which
caused a 404 error when clicking on it, since the website is actually
hosted through the repository.  In other words, clicking on the logo
redirected to "https://myronchen-git.github.io/", but the website is
hosted at
"https://myronchen-git.github.io/software-engineering-portfolio/".