chore: move info specific to me into file
Moved specific information, such as my name, GitHub username, GitHub
profile link, and LinkedIn profile link into an external file,
about.json, in the "portfolio_info" directory, so that it is easier to
modify and keep track of.

The LinkedIn and GitHub links in the NavBar component, the GitHub
username in the "owner" variable in githubApi.js, and the website title
in index.html have been extracted out.  React Helmet has been installed
to dynamically set the website title in App.jsx.